Export and refactor getPagePath function
