+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "golf",
-	"version": "2.2.1",
+	"version": "3.2.3",
 	"author": "RealGolf.Games",
 	"license": "All-Rights-Reserved",
 	"private": true,
