+++ src/lib/components/Nav/Nav.svelte
@@ -1,10 +1,9 @@
 <script lang="ts">
-	import { page } from '$app/stores';
 	import type { User } from '$lib/server/user/types';
-	import { capitalizeFirstLetter } from '$lib/shared/utils/capitalizeFirstLetter';
 	import { faHouse, type IconDefinition } from '@fortawesome/free-solid-svg-icons';
 	import Fa from 'svelte-fa';
 	import Dropdownmenu from './Dropdownmenu.svelte';
+	import Path from './Path.svelte';
 	import Search from './Search.svelte';
 
 	export let auth: string | undefined;
@@ -24,41 +23,6 @@
 			title: 'Home'
 		}
 	];
-
-	$: currentPagePath = $page.url.pathname;
-
-	function getPagePath() {
-		const path = currentPagePath;
-
-		if (path === '/') {
-			return '<strong>Home</strong>';
-		}
-
-		const pathParts = path.split('/').filter((part) => part !== '');
-		let breadcrumbPath = '';
-
-		for (let i = 0; i < pathParts.length; i++) {
-			const pathBefore = pathParts.slice(0, i + 1).join('/');
-			const isLastSegment = i === pathParts.length - 1;
-			const isFirstSegment = i === 0;
-
-			if (isLastSegment) {
-				breadcrumbPath += ` ${
-					isFirstSegment ? '' : "<span style='color: darkgrey;'>/</span>"
-				} <strong>${capitalizeFirstLetter(pathParts[i])}</strong>`;
-			} else {
-				breadcrumbPath += isFirstSegment
-					? ` <a href="/${pathBefore}" style="text-decoration: none;">${capitalizeFirstLetter(
-							pathParts[i]
-					  )}</a>`
-					: ` <span style="color: darkgrey;">/</span> <a href="/${pathBefore}" style="text-decoration: none;">${capitalizeFirstLetter(
-							pathParts[i]
-					  )}</a>`;
-			}
-		}
-
-		return breadcrumbPath;
-	}
 </script>
 
 <nav>
@@ -68,8 +32,7 @@
 				<a href={path} aria-label={title}>
 					<Fa {icon} />
 				</a>
-				<!-- eslint-disable-next-line -->
-				<span class="no-line">{@html getPagePath()}</span>
+				<Path />
 			</li>
 		{/each}
 		<li>
