+++ src/lib/scss/Nav/Nav.scss
@@ -38,11 +38,6 @@ nav {
 			display: flex;
 			flex-direction: row;
 
-			span {
-				margin-left: 15px;
-				white-space: nowrap;
-			}
-
 			.search {
 				margin-right: 0.5rem;
 			}
